refactor: projectCrews/contents 이하 컴포넌트 crewInfo, crewManage로 재분류
