+++ src/features/project/auth/projectCrews/contents/crewInfo/CrewProfile.tsx
@@ -3,9 +3,7 @@
 import Avatar from '@/components/ui/Avatar';
 import PositionBadge from '@/components/ui/badge/PositionBadge';
 import ProjectRoleBadge from '@/components/ui/badge/ProjectRoleBadge';
-import CrewOutButton from '@/features/project/auth/projectCrews/contents/crewOutButtons/CrewOutButton';
 import { getCookie } from 'cookies-next';
-import CrewFwButton from '@/features/project/auth/projectCrews/contents/crewOutButtons/CrewFWButton';
 import TechStackImage from '@/components/ui/TechStackImage';
 import { TechStackItem } from '@/utils/type';
 import TrustGradeBadge from '@/components/ui/badge/TrustGradeBadge';
@@ -13,6 +11,8 @@ import { useCrewDetail } from '@/features/project/auth/projectCrews/service/getC
 import { crewIdState } from '@/features/project/auth/projectCrews/store/CrewIdStateStore';
 import { useRecoilValue } from 'recoil';
 import { bigIntToString } from '@/utils/common';
+import CrewOutButton from '@/features/project/auth/projectCrews/contents/crewManage/CrewOutButton';
+import CrewFwButton from '@/features/project/auth/projectCrews/contents/crewManage/CrewFWButton';
 
 const CrewProfile = () => {
   const crewId = useRecoilValue(crewIdState);
